python-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
python-multipart>=0.0.9
pydantic>=2.8.2
google-genai>=0.2.0
